refactor: modularize tools into individual modules with build-time definitions
- Split monolithic tools.ts into 19 individual tool modules
- Each tool module uses defineTool helper for consistency
- Create build-time generation of toolDefinitions.json
- Update mcp-cloudflare to use generated definitions
- Remove toolDefinitions.ts in favor of build-time generation
- Add comprehensive refactoring documentation

This refactoring improves maintainability by:
- Limiting tool files to ~200 lines for better AI agent context
- Using type-safe defineTool helper pattern
- Generating client-safe definitions without server dependencies
- Following one-tool-per-file architecture
