+++ packages/mcp-server/src/server.ts
@@ -23,8 +23,7 @@ import type {
   ServerRequest,
   ServerNotification,
 } from "@modelcontextprotocol/sdk/types.js";
-import { TOOL_HANDLERS } from "./tools";
-import { TOOL_DEFINITIONS } from "./toolDefinitions";
+import tools from "./tools/index";
 import type { ServerContext } from "./types";
 import { setTag, setUser, startNewTrace, startSpan } from "@sentry/core";
 import { logError } from "./logging";
@@ -148,53 +147,105 @@ export async function configureServer({
   };
 
   for (const resource of RESOURCES) {
-    // Create the handler function once - it's the same for both resource types
-    const resourceHandler = async (
-      url: URL,
-      extra: RequestHandlerExtra<ServerRequest, ServerNotification>,
-    ) => {
-      return await startNewTrace(async () => {
-        return await startSpan(
-          {
-            name: `resources/read ${url.toString()}`,
-            attributes: {
-              "mcp.resource.name": resource.name,
-              "mcp.resource.uri": url.toString(),
-              ...(context.userAgent && {
-                "user_agent.original": context.userAgent,
-              }),
-            },
-          },
-          async () => {
-            if (context.userId) {
-              setUser({
-                id: context.userId,
-              });
-            }
-            if (context.clientId) {
-              setTag("client.id", context.clientId);
-            }
-
-            return resource.handler(url, extra);
-          },
-        );
-      });
-    };
-
     // TODO: this doesnt support any error handling afaict via the spec
-    const source = isTemplateResource(resource)
-      ? resource.template
-      : resource.uri;
+    if (isTemplateResource(resource)) {
+      // Template resource handler receives variables instead of RequestHandlerExtra
+      const templateHandler = async (
+        url: URL,
+        variables: Record<string, string | string[]>,
+      ) => {
+        return await startNewTrace(async () => {
+          return await startSpan(
+            {
+              name: `resources/read ${url.toString()}`,
+              attributes: {
+                "mcp.resource.name": resource.name,
+                "mcp.resource.uri": url.toString(),
+                ...(context.userAgent && {
+                  "user_agent.original": context.userAgent,
+                }),
+              },
+            },
+            async () => {
+              if (context.userId) {
+                setUser({
+                  id: context.userId,
+                });
+              }
+              if (context.clientId) {
+                setTag("client.id", context.clientId);
+              }
 
-    server.registerResource(
-      resource.name,
-      source,
-      {
-        description: resource.description,
-        mimeType: resource.mimeType,
-      },
-      resourceHandler,
-    );
+              // Create a minimal RequestHandlerExtra for the handler
+              const extra: RequestHandlerExtra<
+                ServerRequest,
+                ServerNotification
+              > = {
+                signal: new AbortController().signal,
+                requestId: crypto.randomUUID(),
+                sendNotification: async () => {},
+                sendRequest: async () => ({}) as any,
+              };
+
+              return resource.handler(url, extra);
+            },
+          );
+        });
+      };
+
+      server.registerResource(
+        resource.name,
+        resource.template,
+        {
+          description: resource.description,
+          mimeType: resource.mimeType,
+        },
+        templateHandler,
+      );
+    } else {
+      // Regular resource handler
+      const resourceHandler = async (
+        url: URL,
+        extra: RequestHandlerExtra<ServerRequest, ServerNotification>,
+      ) => {
+        return await startNewTrace(async () => {
+          return await startSpan(
+            {
+              name: `resources/read ${url.toString()}`,
+              attributes: {
+                "mcp.resource.name": resource.name,
+                "mcp.resource.uri": url.toString(),
+                ...(context.userAgent && {
+                  "user_agent.original": context.userAgent,
+                }),
+              },
+            },
+            async () => {
+              if (context.userId) {
+                setUser({
+                  id: context.userId,
+                });
+              }
+              if (context.clientId) {
+                setTag("client.id", context.clientId);
+              }
+
+              return resource.handler(url, extra);
+            },
+          );
+        });
+      };
+
+      server.registerResource(
+        resource.name,
+        resource.uri,
+        {
+          description: resource.description,
+          mimeType: resource.mimeType,
+        },
+        resourceHandler,
+      );
+    }
   }
 
   for (const prompt of PROMPT_DEFINITIONS) {
@@ -261,14 +312,15 @@ export async function configureServer({
     );
   }
 
-  for (const tool of TOOL_DEFINITIONS) {
-    const handler = TOOL_HANDLERS[tool.name];
-
+  for (const [toolKey, tool] of Object.entries(tools)) {
     server.tool(
-      tool.name as string,
+      tool.name,
       tool.description,
-      tool.paramsSchema ? tool.paramsSchema : {},
-      async (...args) => {
+      tool.inputSchema,
+      async (
+        params: any,
+        extra: RequestHandlerExtra<ServerRequest, ServerNotification>,
+      ) => {
         try {
           return await startNewTrace(async () => {
             return await startSpan(
@@ -279,7 +331,7 @@ export async function configureServer({
                   ...(context.userAgent && {
                     "user_agent.original": context.userAgent,
                   }),
-                  ...extractMcpParameters(args[0] || {}),
+                  ...extractMcpParameters(params || {}),
                 },
               },
               async (span) => {
@@ -293,17 +345,15 @@ export async function configureServer({
                 }
 
                 try {
-                  // TODO(dcramer): I'm too dumb to figure this out
-                  // @ts-ignore
-                  const output = await handler(context, ...args);
+                  const output = await tool.handler(params, context);
                   span.setStatus({
                     code: 1, // ok
                   });
                   return {
                     content: [
                       {
-                        type: "text",
-                        text: output,
+                        type: "text" as const,
+                        text: String(output),
                       },
                     ],
                   };
@@ -311,15 +361,7 @@ export async function configureServer({
                   span.setStatus({
                     code: 2, // error
                   });
-                  return {
-                    content: [
-                      {
-                        type: "text",
-                        text: await logAndFormatError(error),
-                      },
-                    ],
-                    isError: true,
-                  };
+                  throw error;
                 }
               },
             );
