+++ packages/mcp-server/src/tools/create-team.ts
@@ -1,0 +1,47 @@
+import { z } from "zod";
+import { setTag } from "@sentry/core";
+import { defineTool } from "./utils/defineTool";
+import { apiServiceFromContext } from "./utils/api-utils";
+import type { ServerContext } from "../types";
+import { ParamOrganizationSlug, ParamRegionUrl } from "../schema";
+
+export default defineTool({
+  name: "create_team",
+  description: [
+    "Create a new team in Sentry.",
+    "",
+    "Be careful when using this tool!",
+    "",
+    "Use this tool when you need to:",
+    "- Create a new team in a Sentry organization",
+    "",
+    "<hints>",
+    "- If any parameter is ambiguous, you should clarify with the user what they meant.",
+    "</hints>",
+  ].join("\n"),
+  inputSchema: {
+    organizationSlug: ParamOrganizationSlug,
+    regionUrl: ParamRegionUrl.optional(),
+    name: z.string().trim().describe("The name of the team to create."),
+  },
+  async handler(params, context: ServerContext) {
+    const apiService = apiServiceFromContext(context, {
+      regionUrl: params.regionUrl,
+    });
+    const organizationSlug = params.organizationSlug;
+
+    setTag("organization.slug", organizationSlug);
+
+    const team = await apiService.createTeam({
+      organizationSlug,
+      name: params.name,
+    });
+    let output = `# New Team in **${organizationSlug}**\n\n`;
+    output += `**ID**: ${team.id}\n`;
+    output += `**Slug**: ${team.slug}\n`;
+    output += `**Name**: ${team.name}\n`;
+    output += "# Using this information\n\n";
+    output += `- You should always inform the user of the Team Slug value.\n`;
+    return output;
+  },
+});
