+++ packages/mcp-server/src/tools/find-tags.ts
@@ -1,0 +1,40 @@
+import { z } from "zod";
+import { setTag } from "@sentry/core";
+import { defineTool } from "./utils/defineTool";
+import { apiServiceFromContext } from "./utils/api-utils";
+import type { ServerContext } from "../types";
+import { ParamOrganizationSlug, ParamRegionUrl } from "../schema";
+
+export default defineTool({
+  name: "find_tags",
+  description: [
+    "Find tags in Sentry.",
+    "",
+    "Use this tool when you need to:",
+    "- Find tags available to use in search queries (such as `find_issues()` or `find_errors()`)",
+  ].join("\n"),
+  inputSchema: {
+    organizationSlug: ParamOrganizationSlug,
+    regionUrl: ParamRegionUrl.optional(),
+  },
+  async handler(params, context: ServerContext) {
+    const apiService = apiServiceFromContext(context, {
+      regionUrl: params.regionUrl,
+    });
+    const organizationSlug = params.organizationSlug;
+
+    setTag("organization.slug", organizationSlug);
+
+    const tagList = await apiService.listTags({ organizationSlug }, {});
+    let output = `# Tags in **${organizationSlug}**\n\n`;
+    if (tagList.length === 0) {
+      output += "No tags found.\n";
+      return output;
+    }
+    output += tagList.map((tag) => [`- ${tag.key}`].join("\n")).join("\n");
+    output += "\n\n";
+    output += "# Using this information\n\n";
+    output += `- You can reference tags in the \`query\` parameter of various tools: \`tagName:tagValue\`.\n`;
+    return output;
+  },
+});
