+++ packages/mcp-server/src/tools/utils/api-utils.ts
@@ -1,0 +1,51 @@
+import { SentryApiService } from "../../api-client/index";
+import { UserInputError } from "../../errors";
+import type { ServerContext } from "../../types";
+
+/**
+ * Create a Sentry API service from server context with optional region override
+ * @param context - Server context containing host and access token
+ * @param opts - Options object containing optional regionUrl override
+ * @returns Configured SentryApiService instance
+ * @throws {UserInputError} When regionUrl is provided but invalid
+ */
+export function apiServiceFromContext(
+  context: ServerContext,
+  opts: { regionUrl?: string } = {},
+) {
+  let host = context.host;
+
+  if (opts.regionUrl?.trim()) {
+    try {
+      const parsedUrl = new URL(opts.regionUrl);
+
+      // Validate that the URL has a proper protocol
+      if (!["http:", "https:"].includes(parsedUrl.protocol)) {
+        throw new UserInputError(
+          `Invalid regionUrl provided: ${opts.regionUrl}. Must include protocol (http:// or https://).`,
+        );
+      }
+
+      // Validate that the host is not just the protocol name
+      if (parsedUrl.host === "https" || parsedUrl.host === "http") {
+        throw new UserInputError(
+          `Invalid regionUrl provided: ${opts.regionUrl}. The host cannot be just a protocol name.`,
+        );
+      }
+
+      host = parsedUrl.host;
+    } catch (error) {
+      if (error instanceof UserInputError) {
+        throw error;
+      }
+      throw new UserInputError(
+        `Invalid regionUrl provided: ${opts.regionUrl}. Must be a valid URL.`,
+      );
+    }
+  }
+
+  return new SentryApiService({
+    host,
+    accessToken: context.accessToken,
+  });
+}
