+++ packages/mcp-server/src/tools/utils/defineTool.ts
@@ -1,0 +1,8 @@
+import type { z } from "zod";
+import type { ToolConfig } from "../types";
+
+export function defineTool<TSchema extends Record<string, z.ZodType>>(
+  config: ToolConfig<TSchema>,
+) {
+  return config;
+}
