+++ packages/mcp-server/src/tools/utils/fetch-utils.ts
@@ -1,0 +1,4 @@
+/**
+ * Re-export of fetch utilities for tool modules.
+ */
+export { fetchWithTimeout } from "../../internal/fetch-utils";
