+++ packages/mcp-server/src/tools/utils/issue-utils.ts
@@ -1,0 +1,10 @@
+/**
+ * Re-export of issue parsing utilities for tool modules.
+ * These utilities handle flexible input formats for Sentry issues.
+ */
+export { parseIssueParams } from "../../internal/issue-helpers";
+
+/**
+ * Re-export of issue formatting utilities for tool modules.
+ */
+export { formatIssueOutput } from "../../internal/formatting";
